Jovian, Data Analysis with Python: Project
As music is a fundamental part of our lives, I was curious to know what makes a song successful on Spotify. I analyzed the Top 200 charts dataset from kaggle by Rakkesh Aravind G. And I passed the Data Analysis with Python: Zero to Pandas course with this project, given it fulfills all the criteria to pass the course.